+++ tools/RAiDER/models/credentials.py
@@ -56,7 +56,7 @@ API_URLS = {'cdsapirc'   : 'https://cds.climate.copernicus.eu/api/v2',
 # api credentials dict
 API_CREDENTIALS_DICT = {
         'cdsapirc' :   {'api' : """\
-                                url: {host}\
+                                \nurl: {host}\
                                 \nkey: {uid}:{key}
                                 """,
                         'help_url' : 'https://cds.climate.copernicus.eu/api-how-to'
@@ -77,6 +77,17 @@ API_CREDENTIALS_DICT = {
                        'help_url': 'https://wiki.earthdata.nasa.gov/display/EL/How+To+Access+Data+With+cURL+And+Wget'
                        }
                 }
+# List existing API variables
+def find_existing_APIenvs():
+    env_api_dict = {}
+    for rc in API_ENVIRONMENTS.keys():
+        uid = API_ENVIRONMENTS [rc]['uid']
+        key = API_ENVIRONMENTS [rc]['key'] 
+        host = API_ENVIRONMENTS [rc]['host']
+        env_api_dict[uid] = os.getenv(uid)
+        env_api_dict[key] = os.getenv(key)
+        env_api_dict[host] = os.getenv(host)
+    return env_api_dict
 
 # Check if API enviroments exists
 def _check_envs(api):
@@ -84,7 +95,7 @@ def _check_envs(api):
     try:
         host = os.getenv(API_ENVIRONMENTS[api]['host'])
     except (RuntimeError):
-        print('Unable to get API url env!')
+        pass
     else:
         host = API_URLS[api] 
 
@@ -103,37 +114,48 @@ def _check_envs(api):
 def check_api(model: str,
               UID: str = None,
               KEY: str = None,
+              output_dir : str = '~/',
               update_flag: bool = False) -> None:
 
-    # Weather model credential filename
+    # Weather model API filename
     # typically stored in home dir as hidden file
     api_filename = API_FILENAME[model]
 
-    # Get API credential from os.env if they exist
-    UID, KEY, URL = _check_envs(api_filename)
+    # Get API credential from os.env if UID/KEY are not inserted
+    if UID is None and KEY is None:
+        UID, KEY, URL = _check_envs(api_filename)
+    else:
+        URL = API_URLS[api_filename]
 
     # skip below if model is HRRR as it does not need API
     if api_filename:    
         # Check if the credential api file exists
-        api_filename_path = os.path.expanduser('~/.') + f'{api_filename}'
+        api_filename_path = os.path.expanduser(output_dir) + f'.{api_filename}'
 
-        # if update flag is on, delete existing file and update it
+        # if update flag is on, overwrite existing file 
         if update_flag:
             if os.path.exists(api_filename_path):
                 os.remove(api_filename_path) 
-            
+        
+        # Check if API_RC file already exists
         if not os.path.exists(api_filename_path): 
+
             # Credential API file does not exist, create it
-            # Need api information
             if UID is None or KEY is None:
                 url = API_CREDENTIALS_DICT[api_filename]['help_url']
-                
-                #Raise ERROR
-                msg = f'{api_filename_path}, API ENVIRONMENTALS and weather model'
-                msg += 'credentials API UID and KEY, do not exist !!'
-                msg += '\nGet API info from ' + '\033[1m' f'{url}' + '\033[0m, and add it!'
-                raise ValueError(msg)
-                
+
+                # Raise ERROR in case only UID or KEY is inserted
+                if UID is not None and KEY is None:
+                    raise ValueError('ERROR: API UID not inserted!')
+                elif UID is None and KEY is not None:
+                    raise ValueError('ERROR: API KEY not inserted!')
+                else:
+                    #Raise ERROR is both UID/KEY are none
+                    raise ValueError(
+                            f'{api_filename_path}, API ENVIRONMENTALS'
+                            f' and API UID and KEY, do not exist !!'
+                            f'\nGet API info from ' + '\033[1m' f'{url}' + '\033[0m, and add it!')
+                    
             # Create file with inputs, do it only once
             print(f'Writing {api_filename_path} locally!')
             with open(api_filename_path, 'w') as f: